(Taint Analysis) Added a taint analysis to genericDataflow
This is a cherry-pick from fbdd8a0b. The taint analysis will
not compile because the genericDataflow API has changed. This will
be fixed in a later commit.

Documentation is in taintAnalysis.h along with lots of notes as
to usability and suitability for binary taint analysis.

Test cases are forthcoming.